refactor[#232] : SignUpByEmail mutation option 이동
